+++ editor_patch/event.h
@@ -75,7 +75,10 @@ enum class AlpineDedEventID : int
     When_Round_Ends,
     Mesh_Animate,
     Mesh_Set_Texture,
-    Mesh_Set_Collision
+    Mesh_Set_Collision,
+    AF_Fullscreen_Image,
+    AF_Fullscreen_Color,
+    Unhide_Glare
 };
 
 constexpr int af_ded_event_to_int(AlpineDedEventID event_id) noexcept